Moved storage module
